Rewrote the entire primary parse step
Made it more direct. Seems to be faster. Did it to be able to properly detect assignment issues.

Also changes the way labels are checked for a bit. Rather than explicitly checking there, it just keeps parsing until it encounters a colon. Have to pass a flag to prevent bogus primaries from being passed as a label, but other than that it's great.